+++ index.js
@@ -1,49 +1,51 @@
 // tooling
-const browserslist  = require('browserslist');
-const jsonToAst     = require('./lib/json-to-ast');
-const normalizeJSON = require('./lib/normalize.json');
-const postcss       = require('postcss');
+const browserslist   = require('browserslist');
+const normalizeRules = require('./lib/normalize.js');
+const postcss        = require('postcss');
 
 // plugin
 module.exports = postcss.plugin('postcss-normalize', (opts) => {
-	const normalizeAST = jsonToAst(normalizeJSON);
-
 	return (root) => {
 		// client browser list
 		const clientBrowserList = browserslist(opts && opts.browsers, {
 			path: root.source && root.source.input && root.source.input.file
 		});
 
-		// required normalize rules
-		const normalizeRules = normalizeAST.nodes.filter(
+		// applied rules from normalize
+		const appliedRules = normalizeRules.map(
 			(rule) => {
-				// required normalize declarations
-				rule.nodes = rule.nodes.filter(
-					(decl) => {
+				const clone = rule.clone();
+
+				clone.nodes = clone.nodes.filter(
+					(decl, index) => {
 						// whether the declaration browser list matches the client browser list
-						const declIsUsed = clientBrowserList.some(
-							(clientBrowser) => browserslist(decl.browserList).some(
+						return clientBrowserList.some(
+							(clientBrowser) => browserslist(rule.nodes[index].browserList).some(
 								(declBrowser) => declBrowser === clientBrowser
 							)
 						);
-
-						// restore source mapping
-						decl.source.input = normalizeJSON.source.input;
-
-						return declIsUsed;
 					}
 				);
 
-				// restore source mapping
-				rule.source.input = normalizeJSON.source.input;
-
-				return rule.nodes.length;
+				return clone;
 			}
-		).map(
-			(rule) => rule.clone()
+		).filter(
+			(rule) => rule.nodes.length
 		);
 
-		// prepend required normalize rules
-		root.prepend(normalizeRules);
+		// check for an @import postcss-normalize insertion point
+		root.walkAtRules(
+			'import',
+			(atrule) => {
+				if (atrule.params === 'postcss-normalize') {
+					atrule.replaceWith(appliedRules);
+				}
+			}
+		);
+
+		if (!appliedRules[0].parent) {
+			// prepend required normalize rules
+			root.prepend(appliedRules);
+		}
 	};
 });