+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Normalize
 
+### 2.1.0 (May 26, 2017)
+
+- Support an insertion point via `@import postcss-normalize`
+- Update tree creation to avoid AST issues with source
+
 ### 2.0.1 (May 21, 2017)
 
 - Update tree creation to avoid AST issues with other PostCSS plugins